code review fixes-syntax correction and moving tests to core package from serverless-plugin
