+++ core/package.json
@@ -4,13 +4,18 @@
   "description": "core framework",
   "main": "alarms.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "tap --coverage-report=html --no-browser --no-check-coverage tests/**/*.test.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "case": "^1.6.3",
     "lodash": "^4.17.21",
-    "yaml": "^2.1.1"
+    "yaml": "^1.10.2"
+  }
+  ,
+  "devDependencies": {
+    "proxyquire": "^2.1.3",
+    "tap": "^15.0.9"
   }
 }
