+++ core/tests/alarms-sqs.test.js
@@ -1,8 +1,8 @@
 'use strict'
 
-const sqsAlarms = require('../../core/alarms-sqs')
+const sqsAlarms = require('../alarms-sqs')
 const { test } = require('tap')
-const defaultConfig = require('../../core/default-config')
+const defaultConfig = require('../default-config')
 const {
   assertCommonAlarmProperties,
   alarmNameToType,
