+++ core/tests/resources/cloudformation-template-stack.json
