+++ core/tests/util.test.js
@@ -7,7 +7,7 @@ const {
   resolveEcsClusterNameAsCfn,
   resolveEcsClusterNameForSub,
   getStatisticName
-} = require('../../core/util')
+} = require('../util')
 
 test('filterObject filters out', (t) => {
   const obj = { a: 1, b: 2, c: 3 }
