+++ package.json
@@ -28,7 +28,8 @@
     "eslint-plugin-standard": "^5.0.0",
     "husky": "^6.0.0",
     "lerna": "^4.0.0",
-    "lint-staged": "^11.0.0"
+    "lint-staged": "^11.0.0",
+    "slic-watch-core": "file:core"
   },
   "engines": {
     "node": ">=12.0"
