+++ serverless-plugin/package.json
@@ -15,7 +15,6 @@
   "devDependencies": {
     "proxyquire": "^2.1.3",
     "serverless": "^3.16.0",
-    "slic-watch-core": "file:../core",
     "tap": "^15.0.9"
   },
   "peerDependencies": {
@@ -29,7 +28,8 @@
     "ajv": "^8.9.0",
     "case": "^1.6.3",
     "lodash": "^4.17.21",
-    "yaml": "^1.10.2"
+    "yaml": "^1.10.2",
+    "slic-watch-core": "file:../core"
   },
   "gitHead": "c4300240f9e854c4eb1c5503b839882ff0cd1cae"
 }
